Utilizar subpath imports do node e realizar pequenos ajustes

+++ backend/src/services/manager.service.ts
@@ -1,8 +1,9 @@
 import { CreationAttributes, Transaction } from "sequelize";
 
-import { Manager, ManagerDTO } from "../interfaces/manager.interface.js";
-import { ManagerModel } from "../models/manager.model.js";
-import { hashPassword } from "../utils/password.utils.js";
+import { Manager, ManagerDTO } from "#interfaces/manager.interface.js";
+import { ManagerModel } from "#models/manager.model.js";
+import { hashPassword } from "#utils/password.utils.js";
+import { isNumeric } from "#utils/string.util.js";
 
 class ManagerService {
   // GET
@@ -30,16 +31,14 @@ class ManagerService {
   }
 
   // POST
-  public async create(userParams: Omit<Manager, "id">, t?: Transaction): Promise<ManagerDTO> {
-    let userParamsNoId: CreationAttributes<Manager> = (({ name, email, password }) => ({
-      name,
-      email,
-      password,
+  public async create(userParams: Manager, t?: Transaction): Promise<ManagerDTO> {
+    let userParamsNoId: CreationAttributes<Manager> = (({ id: _, ...rest }) => ({
+      ...rest,
     }))(userParams);
 
     userParamsNoId.password = hashPassword(userParamsNoId.password);
     const user: Manager = await ManagerModel.create(userParamsNoId, { transaction: t });
-    return { id: user.id, name: user.name, email: user.email };
+    return { id: user.id, name: user.name, email: user.email, origin: user.origin };
   }
 
   // PUT
@@ -72,25 +71,31 @@ class ManagerService {
       transaction: t,
     });
 
-    if (count == 1) {
-      throw new Error("Erro rollback!");
-    }
-
     if (count != 1) {
       throw new Error("Erro no update!");
     }
 
-    return { id: userParamsNoPassword.id, name: userParamsNoPassword.name, email: userParamsNoPassword.email };
+    return {
+      id: userParamsNoPassword.id,
+      name: userParamsNoPassword.name,
+      email: userParamsNoPassword.email,
+      origin: userParamsNoPassword.origin,
+    };
   }
 
-  public async deleteById(id: number, t?: Transaction): Promise<number> {
+  public async deleteById(id: number, t?: Transaction): Promise<string> {
+    if (!isNumeric(id as any)) {
+      throw new Error("O valor de id não é numérico!");
+    }
+
+    const user = await Manager.findByPk(id);
     const count = await ManagerModel.destroy({ where: { id }, transaction: t });
 
-    if (count == 0) {
+    if (count < 1) {
       throw new Error("Usuário não existe");
     }
 
-    return count;
+    return user!.email;
   }
 }
